# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "@openapi:info",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/analog_inputs": {
      "get": {
        "description": "List all Analog Inputs in reef-pi.",
        "tags": [
          "AnalogInput"
        ],
        "summary": "List all Analog Inputs.",
        "operationId": "analogInputList",
        "responses": {
          "200": {
            "description": "analogInput",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/analogInput"
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "Create a new analog input.",
        "tags": [
          "AnalogInput"
        ],
        "summary": "Create an analog input.",
        "operationId": "analogInputCreate",
        "parameters": [
          {
            "description": "The analog input to create",
            "name": "analog input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/analogInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/analog_inputs/{id}": {
      "get": {
        "description": "Get an existing Analog Input.",
        "tags": [
          "AnalogInput"
        ],
        "summary": "Get an Analog Input by id.",
        "operationId": "analogInputGet",
        "parameters": [
          {
            "description": "The Id of the analog input",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/analogInput"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "Update an existing Analog Input.",
        "tags": [
          "AnalogInput"
        ],
        "summary": "Update an Analog Input.",
        "operationId": "analogInputUpdate",
        "parameters": [
          {
            "description": "The Id of the analog input to update",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The analog input to update",
            "name": "analog input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/analogInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "Delete an existing Analog Input.",
        "tags": [
          "AnalogInput"
        ],
        "summary": "Delete an Analog Input.",
        "operationId": "analogInputDelete",
        "parameters": [
          {
            "description": "The Id of the analog input to delete",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/analog_inputs/{id}/read": {
      "post": {
        "description": "Read an Analog Input.",
        "tags": [
          "AnalogInput"
        ],
        "summary": "Read an Analog Input.",
        "operationId": "analogInputRead",
        "parameters": [
          {
            "description": "The Id of the analog input to read",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "analogInput": {
      "type": "object",
      "properties": {
        "driver": {
          "type": "string",
          "x-go-name": "Driver"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "pin": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Pin"
        }
      },
      "x-go-name": "AnalogInput",
      "x-go-package": "github.com/reef-pi/reef-pi/controller/device_manager/connectors"
    },
...
  }
}